move getBlockfromProvider to block, update deps
