+++ packages/statemanager/package.json
@@ -41,15 +41,15 @@
     "tsc": "../../config/cli/ts-compile.sh"
   },
   "dependencies": {
-    "@ethereumjs/block": "^4.0.0",
     "@ethereumjs/common": "^3.0.0",
     "@ethereumjs/rlp": "^4.0.0",
-    "@ethersproject/providers": "^5.6.8",
     "debug": "^4.3.3",
     "ethereum-cryptography": "^1.1.2",
+    "ethers": "^5.7.1",
     "js-sdsl": "^4.1.4"
   },
   "devDependencies": {
+    "@ethereumjs/block": "^4.0.0",
     "@ethereumjs/trie": "^5.0.0",
     "@ethereumjs/util": "^8.0.0",
     "@types/node": "^16.11.7",
